+++ CHANGELOG.rst
@@ -3,9 +3,9 @@ CHANGELOG
 =========
 
 1.0.dev4
-=====
+========
 
-* feature: LinearLearnerEstimator: Add more hyper-parameters
+* spark/pyspark feature: LinearLearnerEstimator: Add more hyper-parameters
 
 
 1.0.3